1. Add scene flags about local coordinate systems and force sign convention (not yet used) 2. Fix many things around L3Geom (most importantly, re-normalization was never done, resulting in instabilities), merge code for different particle shapes in a general function handleSpheresLikeContact 3. Fix incorrect import of gts surface due to LC_NUMERIC 4. Add OpenGL-related stuff to yade-support, which might fix linking issues with natty 5. Move Cell::invTrsf and Cell::refSize to private part, so that it

+++ pkg/dem/Shop.cpp
@@ -577,7 +577,7 @@ Matrix3r Shop::stressTensorOfPeriodicCell(bool smallStrains){
 void Shop::getStressLWForEachBody(vector<Matrix3r>& bStresses, bool revertSign){
 	const shared_ptr<Scene>& scene=Omega::instance().getScene();
 	bStresses.resize(scene->bodies->size());
-	for (unsigned int k=0;k<scene->bodies->size();k++) bStresses[k]=Matrix3r::Zero();
+	for (size_t k=0;k<scene->bodies->size();k++) bStresses[k]=Matrix3r::Zero();
 	FOREACH(const shared_ptr<Interaction>& I, *scene->interactions){
 		if(!I->isReal()) continue;
 		GenericSpheresContact* geom=YADE_CAST<GenericSpheresContact*>(I->geom.get());
